- Added bidirectional one-to-many and many-to-one relationship to user and files tables. Now every registered user can interact only with his files - Extracted handling of duplicate filenames to FilenameDuplicateResolver class

+++ src/main/java/com/dave/dropcube/service/FileServiceImpl.java
@@ -3,8 +3,6 @@ package com.dave.dropcube.service;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
 
@@ -14,6 +12,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.dave.dropcube.dao.FileDAO;
 import com.dave.dropcube.entity.FileEntity;
+import com.dave.dropcube.entity.UserEntity;
+import com.dave.dropcube.exception.ForbiddenFileAccessException;
+import com.dave.dropcube.util.FilenameDuplicationResolver;
 
 @Service
 public class FileServiceImpl implements FileService {
@@ -22,75 +23,19 @@ public class FileServiceImpl implements FileService {
 
 	@Transactional
 	public void uploadFile(FileEntity file) {
-		changeFilenameIfOneIsAlreadyPresent(file);
+		List<FileEntity> allUserFiles = getAllFiles(file.getUserEntity());
+		FilenameDuplicationResolver filenameDuplicationResolver = new FilenameDuplicationResolver();
+		filenameDuplicationResolver.changeFilenameIfOneIsAlreadyPresent(allUserFiles, file);
 		fileDAO.save(file);
 	}
 
-	private void changeFilenameIfOneIsAlreadyPresent(FileEntity file) {
-		List<FileEntity> listOfFiles = getAllFiles(1);
-
-		for (FileEntity tempFile : listOfFiles) {
-			if (tempFile.getName().equals(file.getName())) {
-
-				String newFilename = getFilenameWithNumberSuffixInParenthesis(file
-						.getName());
-				file.setName(newFilename);
-
-				changeFilenameIfOneIsAlreadyPresent(file);
-			}
-		}
-	}
-
-	private String getFilenameWithNumberSuffixInParenthesis(String filename) {
-		String newFilename = null;
-		if (filenameHasAlreadyNumberInParenthesisSuffix(filename)) {
-			newFilename = getNewFilenameWithIncrementedNumberInParenthesisSuffix(filename);
-		} else {
-			newFilename = getNewFilenameWithNumberOneInParenthesisSuffix(filename);
-		}
-		return newFilename;
-	}
-
-	private boolean filenameHasAlreadyNumberInParenthesisSuffix(String filename) {
-		String regexForMatchingFilenameWithNumberInParenthesisSuffix = "(.+)(\\((\\d+)\\))\\..+";
-		return filename
-				.matches(regexForMatchingFilenameWithNumberInParenthesisSuffix);
-	}
-
-	private String getNewFilenameWithIncrementedNumberInParenthesisSuffix(
-			String filename) {
-		String newFilename = null;
-		String regexForGroupingFilenameWithNumberInParenthesisSuffixAndExtension = "(.+)(\\((\\d+)\\))(\\..+)";
-		Pattern pattern = Pattern
-				.compile(regexForGroupingFilenameWithNumberInParenthesisSuffixAndExtension);
-		Matcher matcher = pattern.matcher(filename);
-		if (matcher.find()) {
-			String numberInParenthesis = matcher.group(3);
-			String newNumberInParenthesis = ""
-					+ (Integer.parseInt(numberInParenthesis) + 1);
-
-			newFilename = matcher.group(1) + "(" + newNumberInParenthesis + ")"
-					+ matcher.group(4);
-
-		}
-		return newFilename;
-	}
-
-	private String getNewFilenameWithNumberOneInParenthesisSuffix(
-			String filename) {
-		String newFilename = null;
-		String regexForGroupingFilenameAndExtension = "(.+)(\\..+)";
-		Pattern pattern = Pattern.compile(regexForGroupingFilenameAndExtension);
-		Matcher matcher = pattern.matcher(filename);
-		if (matcher.find()) {
-			newFilename = matcher.group(1) + "(1)" + matcher.group(2);
-		}
-		return newFilename;
-	}
 
 	@Transactional
-	public FileEntity getFile(int fileId) {
-		return fileDAO.getFile(fileId);
+	public FileEntity getFile(UserEntity userEntity, int fileId) {
+		FileEntity file = fileDAO.getFile(fileId);
+		if (!fileMatchesUser(userEntity, file))
+			throw new ForbiddenFileAccessException();
+		return file;
 	}
 
 	/**
@@ -100,12 +45,22 @@ public class FileServiceImpl implements FileService {
 	 * finally wraps bytes from zipped ByteArrayOutputStream with FileEntity
 	 */
 	@Transactional
-	public FileEntity getMultipleFiles(int[] filesIds) throws IOException {
-		ByteArrayOutputStream byteOut = new ByteArrayOutputStream();
+	public FileEntity getMultipleFiles(UserEntity userEntity, int[] filesIds)
+			throws IOException {
 		List<FileEntity> files = fileDAO.getMultipleFiles(filesIds);
-		
+
+		FileEntity file = packMultipleFilesInZipFile(userEntity, files);
+		return file;
+	}
+
+	private FileEntity packMultipleFilesInZipFile(UserEntity user,
+			List<FileEntity> files) throws IOException {
+		ByteArrayOutputStream byteOut = new ByteArrayOutputStream();
 		try (ZipOutputStream out = new ZipOutputStream(byteOut)) {
 			for (FileEntity file : files) {
+				if (!fileMatchesUser(user, file))
+					throw new ForbiddenFileAccessException();
+
 				out.putNextEntry(new ZipEntry(file.getName()));
 				out.write(file.getContent());
 				out.closeEntry();
@@ -113,25 +68,43 @@ public class FileServiceImpl implements FileService {
 		} catch (IOException e) {
 			throw e;
 		}
-		
+
 		FileEntity file = new FileEntity();
 		file.setName("DropCube.zip");
 		file.setContent(byteOut.toByteArray());
 		return file;
 	}
 
+	
+
 	@Transactional
-	public List<FileEntity> getAllFiles(int userId) {
-		return fileDAO.getAllFiles(userId);
+	public List<FileEntity> getAllFiles(UserEntity user) {
+		return fileDAO.getAllFiles(user);
 	}
 
 	@Transactional
-	public void deleteFile(int fileId) {
+	public void deleteFile(UserEntity user, int fileId) {
+		FileEntity file = fileDAO.getFile(fileId);
+		
+		if(!fileMatchesUser(user, file))
+			throw new ForbiddenFileAccessException();
+		
 		fileDAO.deleteFile(fileId);
 	}
 
 	@Transactional
-	public void deleteMultipleFiles(int[] filesIds) {
+	public void deleteMultipleFiles(UserEntity user, int[] filesIds) {
+		for(int fileId : filesIds){
+			FileEntity file = fileDAO.getFile(fileId);
+			if(!fileMatchesUser(user, file))
+				throw new ForbiddenFileAccessException();
+		}
+		
+		
 		fileDAO.deleteMultipleFiles(filesIds);
+	}
+	
+	private boolean fileMatchesUser(UserEntity user, FileEntity file) {
+		return file.getUserEntity().getUserId() == user.getUserId();
 	}
 }
